added admin Controllers and routes

+++ app/controllers/adminController.js
@@ -6,28 +6,28 @@ import resMsg from "../controllers/ErrorsPage.js";
 //signup function for admins
 export  async function admin_signup(req,res) {
     try{
-        const {username, email, password} = req.body;
+        const {username, email, password} = req.body; 
           
-          const EmailExists = await Admin.findOne({email : email});
+          const EmailExists = await Admin.findOne({email : email}); 
           const UsernameExists = await Admin.findOne({username : username});
 
-          if (EmailExists) { 
+          if (EmailExists) { //if the email exitsts
             return res
               .status(400)
               .json({ status: 400, message: "email already exists" });
             }
   
-          if (UsernameExists) { 
+          if (UsernameExists) {  //if the username is already taken 
             return res
               .status(400)
               .json({ status: 400, message: "username already taken" });
             }
-        const hashed_psw = await bcrypt.hash(password,10);
+        const hashed_psw = await bcrypt.hash(password,10); //hashing the password
 
-        await Admin.create({
+        await Admin.create({ //create the admin account by adding those informations in the DB
             username,
             email,
-            hashed_psw,
+            hashed_psw, //the hashed password is saved in the DB 
         });
         return res
          .status(200)
@@ -41,21 +41,20 @@ export  async function admin_signup(req,res) {
 export async function admin_login(req,res){
     try{
         const {email, password} = req.body;
-        const EmailExists = await Admin.findOne({email : email});
-        const PswExists = await Admin.findOne({password : password});
-        
-        if(!EmailExists){
-            return res
-                .status(400)
-                .json({ status: 400, message: "email incorrect!" });
-        }
-        if(!PswExists){
-            return res  
-                .status(400)
-                .json({ status: 400, message: "password incorrect!" });
-        }
-
-        return res
+      
+        const admin = email && (await Admin.findOne({ email: email })); //check if the email exists in the DB
+        if (!admin) //if it doesn't exist
+          return res
+            .status(400)
+            .json({ status: 400, message: "email incorrect!" });
+    
+        const Psw = password && (await bcrypt.compare(password, admin.password)); //compare the password with the hashed password in the DB
+        if (!Psw) //if it's incorrect
+          return res
+            .status(400)
+            .json({ status: 400, message: "password incorrect!" });
+       
+        return res //else : the admin is loged in
         .status(200)
         .json({status: 201, message : "you are loged in"});
             
@@ -67,7 +66,7 @@ export async function admin_login(req,res){
 //delete account function
 export function delete_admin_account(req,res){
     try{
-        await Admin.deleteOne({ _id: req.params.id });
+        await Admin.deleteOne({ _id: req.params.id }); //delete the admin in the DB
     return res
       .status(200)
       .json({ status: 200, data: article, message: "Succesfully deleted" });
@@ -77,8 +76,8 @@ export function delete_admin_account(req,res){
 }
 
 
-
-export function logout(req, res) {
+//log out function for the admins
+export function admin_logout(req, res) {
     try {
       return res
         .status(200)
